code/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.1
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
xgboost==1.7.6
lightgbm==4.0.0
pyyaml==6.0.1
fastapi==0.104.1
pydantic==1.10.13
uvicorn==0.23.2
httpx==0.25.0
pytest==7.4.2
